Move : GradeType 위치 변경 (#17)
